requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.0.0
python-jose==3.3.0
werkzeug==2.3.7
requests==2.31.0
